workin on add post

+++ database/migrations/2020_10_20_051543_add_local_beach_break_to_user_profiles_table.php
@@ -14,7 +14,7 @@ class AddLocalBeachBreakToUserProfilesTable extends Migration
     public function up()
     {
         Schema::table('user_profiles', function (Blueprint $table) {
-            $table->smallInteger('local_beach_break_id')->after('preferred_location')->nullable();
+            $table->integer('local_beach_break_id')->after('preferred_location')->nullable();
         });
     }
